[FE] エラーメッセージの改善
