+++ frontend/app/utils/api-error.ts
@@ -1,0 +1,23 @@
+import { ClientApiError } from "~/types";
+
+export function isClientApiError(e: unknown): e is ClientApiError {
+  if (typeof e === "object" && e !== null) {
+    const typedError = e as {
+      name?: unknown;
+      message?: unknown;
+      stack?: {
+        status?: unknown;
+        payload?: unknown;
+      };
+    };
+    return (
+      typeof typedError.name === "string" &&
+      typeof typedError.message === "string" &&
+      typeof typedError.stack === "object" &&
+      typedError.stack !== null &&
+      typeof typedError.stack.status === "number" &&
+      typeof typedError.stack.payload === "string"
+    );
+  }
+  return false;
+}
